+++ BlockPrinter/Assets/Scripts/RecordDisplay.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Collections;
+using UnityEngine;
+
+namespace BlockPrinter.UserInterface
+{
+
+
+
+    public class RecordDisplay : MonoBehaviour
+    {
+
+        public void UpdateRecord(RecordInfo Record)
+        {
+            
+        }
+    }
+}
